initial bad idea for error messages
This is bad mainly because it pushes fields down the page even as
the user is editing them, but also because it looks like crap,
takes up too much space, makes the whole page ugly and hard to
understand, misses important cases like the case where the
pane width would have to be negative but no single input value is
actually unusable, the design both exacerbates and extends design
weaknesses in how unusuable inputs are detecte and propagated, and
the quality of much of the code in this commit is real bad (e.g.,
what the hell am I doing hiding and unhiding multiple spans when I
need to construct the message string anyway).

I'm committing this because it looks like some of this code might
be adaptable (perhap with a lot of changes) to do something
reasonable.