started work on go-to-top/bottom buttons
There are some problems with this design. The buttons don't appear
in the correct place on mobile. If that's fixed, then they'll be
overlapping text boxes in the form. So mobile browsers need a
separate design for them.

It seems unlikely that this can be fixed for mobile, or a separate
design made to work, without fixing the other bug that, on mobile,
the page height increases (i.e., extra space is added to the end of
the page) when the svg width exceeds the body width. I don't know
an elegant way to fix that, and it appears mobile-specific.

As a possibly separate matter, pressing the buttons doesn't even
scroll on mobile browsers (at least as simulated in responsive
design mode in Firefox and its analogue in Chrome).

Other problems with these buttons include that:

While scrolling is smooth on Firefox, for some reason it is not
smooth on Chromium-based browsers. That would need to be fixed.

Mousedown on a button properly makes it look pressed, but if the
mouse cursor is moved off the button before mouseup, the button
shows no change in appearance until mouseup. This is arguably not
objectively a bug, but it differs from what users probably expect,
since default-styled buttons on most (if not all) platforms don't
behave this way. Even so, that might not really need to be fixed.

I'm not sure I'm going to use any of what this commit introduces,
becuase it has so many problems.